Fix update style gui
